Split checkDistance service to separate classes
Move logic with Vibration to VibratorService. Move Notifications logic to separate Context. Create a skeleton class for CheckDistance (Camera) Service
