+++ app/src/main/AndroidManifest.xml
@@ -1,6 +1,6 @@
 <?xml version="1.0" encoding="utf-8"?>
 <manifest xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools">
+    xmlns:tools="http://schemas.android.com/tools" >
 
     <uses-permission android:name="android.permission.CAMERA" />
 
@@ -27,7 +27,12 @@
         android:theme="@style/Theme.SafeDistance"
         tools:targetApi="31">
         <service
-            android:name=".CheckDistance"
+            android:name=".service.CheckDistanceService"
+            android:foregroundServiceType="camera"
+            android:enabled="true"
+            android:exported="false" />
+        <service
+            android:name=".service.VibratorService"
             android:foregroundServiceType="camera"
             android:enabled="true"
             android:exported="false" />
